fix: resolve all TypeScript build errors
- Move Message, MessagePage, ThreadSummary types to channels.ts (canonical
  home) and re-export from threads.ts; serverStore.ts imports them from there
- Add membershipMode: string to ServerStore interface and initializer
- Fix THREAD_CREATE gateway event cast (parent_message_id -> message_id)
- Replace Array.at(-1) with index access (ES2020 lib target)
- Widen ChannelSidebar status prop to include "reconnecting"
- Add missing is_bot field to MemberContextMenu test fixture
- Add missing reactions/thread fields to Message fixtures in tests
- Fix caughtError type narrowing in useAccountManager test
- Remove invalid Parameters<setState> cast in serverStore test
